+++ RigidBodies/PhysicsProject/SolarSystemParser.cpp
@@ -31,16 +31,16 @@ void SolarSystemParser::ParsePlanetData(XMLElement* dataElement)
 	planetData.Name = std::string(it->GetText());
 
 	it = it->NextSiblingElement("mass");
-	planetData.Mass = atof(it->GetText());
+	planetData.Mass = static_cast<float>(atof(it->GetText()));
 
 	it = it->NextSiblingElement("radius");
-	planetData.Radius = atof(it->GetText());
+	planetData.Radius = static_cast<float>(atof(it->GetText()));
 
 	it = it->NextSiblingElement("distance");
-	planetData.Distance = atof(it->GetText());
+	planetData.Distance = static_cast<float>(atof(it->GetText()));
 
 	it = it->NextSiblingElement("velocity");
-	planetData.Velocity = atof(it->GetText());
+	planetData.Velocity = static_cast<float>(atof(it->GetText()));
 
 	it = it->NextSiblingElement("color");
 	float r = it->FloatAttribute("r");
